Probably should not commit without testing, but such are things this week.
Trial change in echo top height so it is set to 0 rather than SPVAL before
an attempt is made to give it a valid definition.  As points without > 18.3 dBZ
reflectivity in the column are not redefined, the initial value has consequences.

Hope is that it will correct a copygb2 issue with the field.